Move loggers to loggers.py file
This move allows us to stop tracking tests for these two objects and
have the package's coverage report more closely representing what
it should.
