+++ bayes_opt/__init__.py
@@ -1,6 +1,6 @@
 from .bayesian_optimization import BayesianOptimization, Events
 from .util import UtilityFunction
-from .observer import ScreenLogger, JSONLogger
+from .logger import ScreenLogger, JSONLogger
 
 __all__ = [
     "BayesianOptimization",
